app/requirements.txt
Django
psycopg2
python-dotenv
djangorestframework
django-cors-headers
drf-yasg